update document field type

+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/implementation/util/Transforms.java
@@ -418,10 +418,10 @@ public class Transforms {
             DocumentFieldHelper.setValue(documentField, innerDocumentField.getValuePhoneNumber());
         } else if (com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentFieldType.NUMBER.equals(
             innerDocumentField.getType())) {
-            DocumentFieldHelper.setValue(documentField, innerDocumentField.getValueNumber());
+            DocumentFieldHelper.setValue(documentField, innerDocumentField.getValueNumber().doubleValue());
         } else if (com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentFieldType.INTEGER.equals(
             innerDocumentField.getType())) {
-            DocumentFieldHelper.setValue(documentField, innerDocumentField.getValueInteger());
+            DocumentFieldHelper.setValue(documentField, innerDocumentField.getValueInteger().longValue());
         } else if (com.azure.ai.formrecognizer.documentanalysis.implementation.models.DocumentFieldType.SELECTION_MARK.equals(
             innerDocumentField.getType())) {
             if (innerDocumentField.getValueSelectionMark() == null) {
